requirements.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.7.1
django-redis==5.2.0
djangorestframework==3.13.1
dnspython==2.2.0
et-xmlfile==1.1.0
eventlet==0.33.0
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.10.0
kombu==5.2.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.0
prompt-toolkit==3.0.27
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0